Update Package structure documentation. Move ``dali.interfaces`` to ``dali.drivers.daliserver``. Add ``dali.exceptions`` module and move all DALI related exceptions to it.
