+++ CHANGES.rst
@@ -1,17 +1,32 @@
 Changes
 =======
 
+0.5 (unreleased)
+----------------
+
+- Add ``dali.exceptions`` module and move all DALI related exceptions to it.
+  [rnix]
+
+- Move ``dali.interfaces`` to ``dali.drivers.daliserver``.
+  [rnix]
+
+- Update Package structure documentation.
+  [rnix]
+
+
 0.4
 ---
 
 - PyPI support
   [dgomes]
 
+
 0.3
 ---
 
 - Python 3 support
   [sde1000]
+
 
 0.2
 ---
@@ -21,6 +36,7 @@ Changes
 
 - Basic implementation of LUNATONE Dali USB driver.
   [rnix]
+
 
 0.1
 ---
